doap: update
Thanks Jordi for your past contributions!

The homepage can be the same as the main gedit one, which contains a
link to a list of gedit plugins (both core ones and these ones).

The mailing list has been archived.

For the bug-database, we want good bug reports and avoiding getting many
duplicates.