Remove findinfiles plugin
The plugin doesn't have any maintainer.

When merged, we asked the author to continue maintaining it, but free
software doesn't work this way. (BTW the plugin initial work was funded,
but reviews and maintenance wasn't).

Also, since it's written in Vala it's harder for us to work with. We
already have C, Python and a little of Objective C.

If somebody wants to maintain it, an easy solution is to put it in its
own repo.

A replacement is to use an external tool, like grep / git grep.